Fix issue with token services
The token services already have a token parameter appended at the end of the url. agsLegend appends more parameters to use jsonp request to the url later on, and this screws up the request URL. This changes the url handling to make sure that the url's parameters are preserved and appended correctly.